requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cors==1.0.1
cycler==0.12.1
dask==2024.6.0
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
executing==2.0.1
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.5.4
gevent==24.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pymongo==4.7.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tldextract==5.1.2
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
